moved all preferences models to the separate folder
