+++ src/appshell/appshellmodule.cpp
@@ -36,20 +36,20 @@
 #include "view/appmenumodel.h"
 #include "view/notationpagemodel.h"
 #include "view/aboutmodel.h"
-#include "view/preferencesmodel.h"
-#include "view/generalpreferencesmodel.h"
-#include "view/updatepreferencesmodel.h"
-#include "view/appearancepreferencesmodel.h"
-#include "view/programmestartpreferencesmodel.h"
-#include "view/folderspreferencesmodel.h"
-#include "view/noteinputpreferencesmodel.h"
-#include "view/advancedpreferencesmodel.h"
-#include "view/canvaspreferencesmodel.h"
-#include "view/scorepreferencesmodel.h"
-#include "view/importpreferencesmodel.h"
-#include "view/iopreferencesmodel.h"
 #include "view/startupmodel.h"
-#include "view/commonaudioapiconfigurationmodel.h"
+#include "view/preferences/preferencesmodel.h"
+#include "view/preferences/generalpreferencesmodel.h"
+#include "view/preferences/updatepreferencesmodel.h"
+#include "view/preferences/appearancepreferencesmodel.h"
+#include "view/preferences/programmestartpreferencesmodel.h"
+#include "view/preferences/folderspreferencesmodel.h"
+#include "view/preferences/noteinputpreferencesmodel.h"
+#include "view/preferences/advancedpreferencesmodel.h"
+#include "view/preferences/canvaspreferencesmodel.h"
+#include "view/preferences/scorepreferencesmodel.h"
+#include "view/preferences/importpreferencesmodel.h"
+#include "view/preferences/iopreferencesmodel.h"
+#include "view/preferences/commonaudioapiconfigurationmodel.h"
 
 using namespace mu::appshell;
 using namespace mu::framework;
