+++ src/appshell/view/preferences/advancedpreferencesmodel.cpp
