+++ src/appshell/view/preferences/advancedpreferencesmodel.h
