+++ src/appshell/view/preferences/canvaspreferencesmodel.cpp
