+++ src/appshell/view/preferences/canvaspreferencesmodel.h
