+++ src/appshell/view/preferences/folderspreferencesmodel.cpp
