+++ src/appshell/view/preferences/folderspreferencesmodel.h
