+++ src/appshell/view/preferences/generalpreferencesmodel.cpp
