+++ src/appshell/view/preferences/generalpreferencesmodel.h
