+++ src/appshell/view/preferences/noteinputpreferencesmodel.cpp
