+++ src/appshell/view/preferences/preferencesmodel.cpp
