+++ src/appshell/view/preferences/scorepreferencesmodel.cpp
