+++ src/appshell/view/preferences/scorepreferencesmodel.h
